Move path suffix tests into their module
